move bbox_overlap to utils.bbox
