+++ chainercv/utils/__init__.py
@@ -1,3 +1,4 @@
+from chainercv.utils.bbox.bbox_overlap import bbox_overlap  # NOQA
 from chainercv.utils.download import cached_download  # NOQA
 from chainercv.utils.download import extractall  # NOQA
 from chainercv.utils.extension import check_type  # NOQA
