+++ docs/source/reference/links.rst
@@ -2,14 +2,3 @@ Links
 =====
 
 .. module:: chainercv.links
-
-
-Utils
------
-
-bbox
-~~~~
-
-bbox_overlap
-""""""""""""
-.. autofunction:: chainercv.links.utils.bbox_overlap
